feat(ctrPersonCard): load person by NationalNo & fix Gender type - Added LoadPersonInfo(NationalNo) to load person using National Number - Updated Gender type to byte in clsPerson and data access for consistency

+++ DVLD_(Driving License Management System)/DVLD_BuisnessDVLD_Buisness/clsPerson.cs
@@ -138,7 +138,7 @@ namespace DVLD_BuisnessDVLD_Buisness
 
             DateTime DateOfBirth = DateTime.Now;
             int PersonID = -1, NationalityCountryID = -1;
-            short Gendor = 0;
+            byte Gendor = 0;
 
             bool IsFound = clsPersonData.GetPersonInfoByNationalNo
                                 (
